+++ HISTORY.md
@@ -3,7 +3,7 @@
 ## 1.1.0 / 2021-08-09
 
 - Added stored value retrieval
-- Added URLencode() wrappers on each URL portion
+- Added URLencode() wrapper to data view URL
 
 ## 1.0.2 / 2021-07-13